iterate over ids instead of loading all ids into memory
