+++ corehq/apps/domain/management/commands/delete_domain.py
@@ -1,16 +1,8 @@
 import textwrap
 
 from django.core.management.base import BaseCommand
-from django.db.models import Q
-
 
 from corehq.apps.domain.dbaccessors import iter_all_domains_and_deleted_domains_with_name
-from corehq.apps.domain.utils import silence_during_tests
-from corehq.sql_db.util import (
-    estimate_partitioned_row_count,
-    paginate_query_across_partitioned_databases,
-)
-from corehq.util.log import with_progress_bar
 
 
 class Command(BaseCommand):
@@ -55,21 +47,3 @@ class Command(BaseCommand):
             assert domain_obj.name == domain_name  # Just to be really sure!
             domain_obj.delete(leave_tombstone=True)
         print("Operation completed")
-
-
-def iter_ids(model_class, field, domain, chunk_size=1000):
-    where = Q(domain=domain)
-    rows = paginate_query_across_partitioned_databases(
-        model_class,
-        where,
-        values=[field],
-        load_source='delete_domain',
-        query_size=chunk_size,
-    )
-    yield from with_progress_bar(
-        (r[0] for r in rows),
-        estimate_partitioned_row_count(model_class, where),
-        prefix="",
-        oneline="concise",
-        stream=silence_during_tests(),
-    )
